refactor: move brand delivery stats from campaigns.ts to brand.ts
GET /v1/brands/:id/stats calls email-gateway, not campaign-service — it
doesn't belong in the campaigns router. Moved to brand.ts alongside
the other brand stats routes (/brands/stats/costs, /brands/:id/stats/costs).

Also extracted fetchDeliveryStats to src/lib/delivery-stats.ts since it's
shared between campaigns.ts (campaign stats + SSE stream) and brand.ts.

Breaking change: path param renamed from {brandId} to {id} for consistency
with all other brand routes.
